Create Policy instructions
Running Adobe Remote Update Manager can take a few minutes and is not good to run in an Extension Attribute.  Instead, make a policy to figure out the count.

A bug in this is sometimes it will report back 2 when there really is 0.  

Now make a smart group that reads the value of the EA and then you can run sdrum when there are updates available.